Partially implement GCircuit with minimal GBlock implementation
Adds partial implementation of GCircuit that is able to render dummy
GBlocks with minimal GBlock implementation that is enough to provide
these block with basic functionality. Includes lots of TODOs.

Adds assets folder to provide mock GBlock texture.
WARNING: This folder is not properly added to build system and currently
should not be used as directory to place project's assets.