Switch to Flix01/imgui that has file picker addon
An attempt to switch to Flix01/imgui was done on #19 but where were
issues with building it on Windows. This commit is result of this
investigation.

Issue was caused by not including windows.h before other windows
libraries such as windef.h by addons from Flix01/imgui.
I don't know if it's our build config's issue or with Flix01/imgui
but I applied the fastest (though not the most elegant) solution.

NOTE: When updating ImGui, manual fix needs to be performed:
In imgui/addons/imguifilesystem.h, before including windef.h,
windows.h should be included.

Closes #51. Fixes #19.